finish qr code share
