+++ app/src/main/java/com/github/palFinderTeam/palfinder/ProfileActivity.kt
@@ -21,6 +21,7 @@ import com.github.palFinderTeam.palfinder.meetups.activities.MeetUpView
 import com.github.palFinderTeam.palfinder.profile.ProfileUser
 import com.github.palFinderTeam.palfinder.profile.USER_ID
 import com.github.palFinderTeam.palfinder.utils.Response
+import com.github.palFinderTeam.palfinder.utils.image.QRCode
 import com.google.zxing.BarcodeFormat
 import com.google.zxing.integration.android.IntentIntegrator
 import com.journeyapps.barcodescanner.BarcodeEncoder
@@ -193,14 +194,7 @@ class ProfileActivity : AppCompatActivity() {
         //Encode text in editText into QRCode image into the specified size using barcodeEncoder
         val bitmap = barcodeEncoder.encodeBitmap(USER_ID+intent.getStringExtra(USER_ID), BarcodeFormat.QR_CODE, resources.getInteger(R.integer.QR_size), resources.getInteger(R.integer.QR_size))
 
-        //Set up the popup image
-        val imagePopup = ImagePopup(this)
-        //Convert the bitmap(QR Code) into a drawable
-        val d: Drawable = BitmapDrawable(resources, bitmap)
-
-        //Displays the popup image
-        imagePopup.initiatePopup(d);
-        imagePopup.viewPopup()
+        QRCode.shareQRcode(bitmap, this)
 
     }
